Adds test_objects and test_conditions
